# docs/swagger.yaml
basePath: /api/v1
definitions:
  controller.InvoiceViewFormat:
    properties:
      invoice_id:
        type: string
      order_details: {}
      order_id:
        type: string
      payment_due: {}
      payment_due_date:
        type: string
      payment_method:
        type: string
      payment_status:
        type: string
      table_number: {}
    type: object
...
  model.FoodModel:
    properties:
      created_at:
        type: string
      food_id:
        type: string
      food_image:
        type: string
      id:
        type: string
      menu_id:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      price:
        type: number
      updated_at:
        type: string
    required:
    - food_image
    - menu_id
    - name
    - price
    type: object
  model.InvoiceModel:
    properties:
      created_at:
        type: string
      id:
        type: string
      invoice_id:
        type: string
      order_id:
        type: string
      payment_due_date:
        type: string
      payment_method:
        type: string
      payment_status:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:3000
info:
  contact:
    name: API Support
    url: http://www.swagger.io/support
  description: A comprehensive Restaurant Management System API for managing users,
    menus, food items, orders, tables, invoices, and notes. Available on both local
    and online servers.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Restaurant Management API
  version: "1.0"
  x-servers:
  - description: Local Development Server
    url: http://localhost:3000/api/v1
  - description: Online Production Server
    url: https://restaurant-management-f9kx.onrender.com/api/v1
paths:
  /food:
    post:
      consumes:
      - application/json
      description: Create a food item in the restaurant system
      parameters:
      - description: Food data
        in: body
        name: food
        required: true
        schema:
          $ref: '#/definitions/model.FoodModel'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.FoodModel'
      summary: Create a new food item
      tags:
      - Food
  /food-update/{id}:
    put:
      consumes:
      - application/json
      description: Update food item details by ID
      parameters:
      - description: Food ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated food data
        in: body
        name: food
        required: true
        schema:
          $ref: '#/definitions/model.FoodModel'
      produces:
      - application/json
      responses:
        "200":
          description: Food updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Food not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a food item
      tags:
      - Food
  /food/{id}:
    get:
      description: Retrieve a food item by its ID
      parameters:
      - description: Food ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Food details
          schema:
            $ref: '#/definitions/model.FoodModel'
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a specific food item
      tags:
      - Food
  /foods:
    get:
      description: Retrieve a paginated list of all food items
      parameters:
      - description: 'Number of records per page (default: 10)'
        in: query
        name: recordPerPage
        type: integer
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of food items with pagination
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all food items with pagination
      tags:
      - Food
  /invoice:
    post:
      consumes:
      - application/json
      description: Create a new invoice for an order with payment status and method
      parameters:
      - description: Invoice data (order_id required)
        in: body
        name: invoice
        required: true
        schema:
          $ref: '#/definitions/model.InvoiceModel'
      produces:
      - application/json
      responses:
        "200":
          description: Invoice created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request - validation error
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new invoice
      tags:
      - Invoice
  /invoice/{id}:
    delete:
      description: Delete an invoice by invoice ID
      parameters:
      - description: Invoice ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Invoice deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete an invoice
      tags:
      - Invoice
    get:
      description: Retrieve invoice details with order items by invoice ID
      parameters:
      - description: Invoice ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Invoice details with order information
          schema:
            $ref: '#/definitions/controller.InvoiceViewFormat'
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a specific invoice
      tags:
      - Invoice
    put:
      consumes:
      - application/json
      description: Update invoice payment status and payment method by invoice ID
      parameters:
      - description: Invoice ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated invoice data (payment_method and payment_status)
        in: body
        name: invoice
        required: true
        schema:
          $ref: '#/definitions/model.InvoiceModel'
      produces:
      - application/json
      responses:
        "200":
          description: Invoice updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update an invoice
      tags:
      - Invoice
  /invoices:
    get:
      description: Retrieve a list of all invoices
      produces:
      - application/json
      responses:
        "200":
          description: List of invoices
          schema:
            items:
              $ref: '#/definitions/model.InvoiceModel'
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all invoices
      tags:
      - Invoice
...
schemes:
- http
- https
swagger: "2.0"